+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add "Print Using GB Printer" event to send either the current Background or Overlay to a connected GB Printer. Game must be run on real hardware with a printer connected for this feature to work (based on "Print Screen Background" plugin by [@pau-tomas](https://github.com/pau-tomas))
 - Add "Set Camera Property" event for changing camera deadzone and offset values
 - Add "Actor Effects" event for playing effect animations on actors (based on "Actor FX" plugin by [@pau-tomas](https://github.com/pau-tomas))
+- Add "Thread Start" and "Stop Thread" events allowing running scripts in the background.
 
 ### Changed